% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmify.R
\docType{package}
\name{ppmify-package}
\alias{ppmify-package}
\title{ppmify package}
\description{
ppmify is a micropackage to help fitting Poisson point process
  models to point data, using one of the very many types Poisson regression
  models available in R. The main function is \code{ppmify}, which takes some
  coordinates, an area of interest and covariates and produces a \code{ppm}
  object, which can be used in place of a dataframe in a Poisson regression
  model. See \code{\link{ppmify}} for details.
}

