% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmify.R
\name{ppmify}
\alias{ppmify}
\title{create a ppm object from point data}
\usage{
ppmify(coords, area = NULL, covariates = NULL, method = c("grid",
  "count"))
}
\arguments{
\item{coords}{a matrix, dataframe or SpatialPoints* object giving the
coordinates of the points to use in the PPM analysis}

\item{area}{an optional extent, SpatialPolygons* or Raster* object giving the
area over which to model the point process. If ignored, a rectangle
defining the extent of \code{coords} will be used instead.}

\item{covariates}{an optional Raster* object containing covariates for
modelling the poit process}

\item{method}{the method for defining integration area. This will either
generate a set of integration points with appropriate weights, or count the
number of points falling in each cell (if \code{method = 'full'}). See \code{Details} for information on the available approaches.}
}
\details{
<integration details to be added>
}

